File strcture changed
