+++ src/libs/atoms/DropdownJ/DropdownJ.js
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import { View, StyleSheet, Text } from 'react-native';
 import { Picker } from '@react-native-picker/picker';
+import DropdownJstyles from './DropdownJ.styles';
 
 const generateRange = (start, end) => {
   const range = [];
@@ -33,11 +34,11 @@ const DropdownJ = () => {
   const [selectedYear, setSelectedYear] = useState(years[0]);
 
   return (
-    <View style={styles.container}>
-      <View style={styles.input}> 
+    <View style={DropdownJstyles.container}>
+      <View style={DropdownJstyles.input}> 
         <Picker
           selectedValue={selectedDay}
-          style={styles.picker}
+          style={DropdownJstyles.picker}
           onValueChange={(itemValue) => setSelectedDay(itemValue)}
         >
           {days.map((day) => (
@@ -46,10 +47,10 @@ const DropdownJ = () => {
           ))}
         </Picker>
       </View>
-      <View style={styles.input}> 
+      <View style={DropdownJstyles.input}> 
       <Picker
         selectedValue={selectedMonth}
-        style={styles.picker}
+        style={DropdownJstyles.picker}
         onValueChange={(itemValue) => setSelectedMonth(itemValue)}
       >
         {months.map((month) => (
@@ -57,10 +58,10 @@ const DropdownJ = () => {
         ))}
       </Picker>
       </View>
-      <View style={styles.input}> 
+      <View style={DropdownJstyles.input}> 
       <Picker
         selectedValue={selectedYear}
-        style={styles.picker}
+        style={DropdownJstyles.picker}
         onValueChange={(itemValue) => setSelectedYear(itemValue)}
       >
         {years.map((year) => (
@@ -71,45 +72,5 @@ const DropdownJ = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-
-  },
-  picker: {
-    height: 50,
-    width: 90,
-    color: '#e5e4e2',
-  },
-  appContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-
-  },
-  title: {
-    fontSize: 18,
-    marginBottom: 10,
-    
-  },
-  input: {
-    flexDirection: 'row',
-    borderWidth: 1,
-    borderRadius: 20,
-    marginBottom: 25,
-    paddingHorizontal: 10,
-    borderColor: '#e5e4e2',
-    alignItems: 'center',
-    height: 60,
-    width: 90,
-    marginLeft: 5,
-  }
-});
-
-
-
 
 export default DropdownJ;
