+++ src/libs/atoms/FileUploadJ/FileUploadJ.styles.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+import { Color, Padding } from "../../../../GlobalStyles";
+
+const FileUploadJstyles = StyleSheet.create({
+  image: {
+    width: 200,
+    height: 200,
+    marginTop: 20,
+    marginBottom: 20,
+  },
+});
+
+
+export default FileUploadJstyles;
