+++ src/libs/atoms/InputJ/InputJ.styles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+import { Color, Padding } from "../../../../GlobalStyles";
+
+const InputJstyles = StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: 'center',
+        padding: 20,
+      },
+      
+      
+      
+});
+
+
+export default InputJstyles;
