+++ src/libs/atoms/PasswordInputJ/PasswordInputJ.styles.js
@@ -1,0 +1,32 @@
+import { StyleSheet } from 'react-native';
+import { Color, Padding } from "../../../../GlobalStyles";
+
+const PasswordInputJstyles = StyleSheet.create({
+    // container: {
+    //     flex: 1,
+    //     justifyContent: 'center',
+    //     padding: 20,
+    //   },
+     
+    //   inputContainer: {
+    //     flexDirection: 'row',
+    //     borderWidth: 1,
+    //     borderRadius: 20,
+    //     marginBottom: 25,
+    //     paddingHorizontal: 10,
+    //     borderColor: '#e5e4e2',
+    //     alignItems: 'center',
+    //     height: 50,
+    //     width: 280,
+    //   },
+      
+    //   icon: {
+    //     marginRight: 10,
+    //   },
+      
+      
+      
+});
+
+
+export default PasswordInputJstyles;
