+++ src/libs/screenComponents/sware-home-component/organisms/FormScreenJ/FormScreenJ.js
@@ -1,0 +1,34 @@
+import React from 'react';
+import { View, StyleSheet, Text } from 'react-native';
+import ButtonJ from '../../../../atoms/ButtonJ/ButtonJ';
+import CheckboxJ from '../../../../atoms/CheckboxJ/CheckboxJ';
+import DropdownJ from '../../../../atoms/DropdownJ/DropdownJ';
+import InputJ from '../../../../atoms/InputJ/InputJ';
+import PasswordInputJ from '../../../../atoms/PasswordInputJ/PasswordInputJ';
+import { Border, Color } from "../../../../../../GlobalStyles"
+import CssJ from '../../../../atoms/CommonCss/CssJ';
+import FormScreenJstyles from './FormScreenJ.styles';
+
+const FormScreenJ = () => {
+    return (
+        <View style={CssJ.container}>
+                <View style={FormScreenJstyles.boxInput}>
+                <InputJ
+                iconName = "fax"
+                inputText = "CARD NUMBER" />
+                <InputJ
+                iconName = "person"
+                inputText = "CARDHOLDER'S NAME" />
+                <InputJ
+                iconName = "call"
+                inputText = "CONTACT NO" />
+                <PasswordInputJ/>
+                <DropdownJ />
+                <CheckboxJ/>
+                <ButtonJ/>
+                </View>
+        </View>
+    );
+};
+
+export default FormScreenJ;
